requirements.txt
alembic
cryptography
fastapi
passlib
PyJWT
PyMySQL
python-dotenv
requests
sentence-transformers
SQLAlchemy
uvicorn